fix: use new procedure names
